+++ Scripts/Editor/TMP_PostBuildProcessHandler.cs
@@ -25,7 +25,7 @@ namespace TMPro
                 {
                     file = Path.Combine(pathToBuiltProject, "Classes/UI/Keyboard.mm");
                 }
-                
+
                 string content = File.ReadAllText(file);
                 content = content.Replace("FILTER_EMOJIS_IOS_KEYBOARD 1", "FILTER_EMOJIS_IOS_KEYBOARD 0");
                 File.WriteAllText(file, content);